Factor out the definitions to their own module
